+++ src/modules/core/webviews/table/app.ts
@@ -247,8 +247,16 @@ function getHtmlBody(connectionConfig: PoolConnectionConfig): string {
     <div id="table-header"></div>
     <div id="table"></div>
     <div id="table-selection-context-menu" class="select-options">
-        <div class="select-option" id="table-selection-context-menu-delete-rows">Delete Rows</div>
-        <div class="select-option" id="table-selection-context-menu-delete-rows-clear">Clear: Delete Rows</div>
+        <div class="select-option" id="table-selection-context-menu-set-null">Set NULL</div>
+        <div class="select-option" id="table-selection-context-menu-delete-row">Delete Row</div>
+        <div class="select-option" id="table-selection-context-menu-delete-row-clear">Clear: Delete Row</div>
+        <div class="horizontal-divider"></div>
+        <div class="select-option" id="table-selection-context-menu-delete-rows">Selected Delete Rows</div>
+        <div class="select-option" id="table-selection-context-menu-delete-rows-clear">Clear: Selected Delete Rows</div>
+        <div class="horizontal-divider"></div>
+        <div class="select-option" id="table-selection-context-menu-select-row">Select Row</div>
+        <div class="select-option" id="table-selection-context-menu-select-column">Select Column</div>
+        <div class="select-option" id="table-selection-context-menu-select-all">Select All</div>
     </div>
 `;
 }