+++ src/modules/core/webviews/table/webview/tableSelectionMode.ts
@@ -1,8 +1,9 @@
 import { copyToClipboard } from "../../webview-helper/copyToClipboard";
+import { logError } from "../../webview-helper/logger";
 import { selectTextInContentEditableDiv } from "../../webview-helper/selectTextInContentEditableDiv";
 import { exportDataSeparatorDefaults } from "./exportData";
 import { updateOnPushChangesState } from "./push";
-import { getQueryResultDeletions, setQueryResultDeletions } from "./query";
+import { getQueryResult, getQueryResultDeletions, setQueryResultDeletions, updateDataChanges } from "./query";
 import { getTableElement } from "./table";
 
 let isSelectionModeMouseDown = false;
@@ -14,7 +15,17 @@ let colWidth = 180;
 const selectionModeElement = document.getElementById('selection-mode') as HTMLElement;
 const tableSelectionContextMenuDeleteRowsElement = document.getElementById('table-selection-context-menu-delete-rows') as HTMLElement;
 const tableSelectionContextMenuDeleteRowsClearElement = document.getElementById('table-selection-context-menu-delete-rows-clear') as HTMLElement;
+const tableSelectionContextMenuDeleteRowElement = document.getElementById('table-selection-context-menu-delete-row') as HTMLElement;
+const tableSelectionContextMenuDeleteRowClearElement = document.getElementById('table-selection-context-menu-delete-row-clear') as HTMLElement;
+const tableSelectionContextMenuSelectRowElement = document.getElementById('table-selection-context-menu-select-row') as HTMLElement;
+const tableSelectionContextMenuSelectColumnElement = document.getElementById('table-selection-context-menu-select-column') as HTMLElement;
+const tableSelectionContextMenuSelectAllElement = document.getElementById('table-selection-context-menu-select-all') as HTMLElement;
+const tableSelectionContextMenuSetNullElement = document.getElementById('table-selection-context-menu-set-null') as HTMLElement;
 const tableSelectionContextMenuElement = document.getElementById('table-selection-context-menu') as HTMLElement;
+let contextMenuTopLeftPosition = {
+    row: 0,
+    col: 0
+};
 
 export function addSelectionModeOverlayEventListeners(overlayElement: HTMLElement | null) {
     if (overlayElement) {
@@ -83,11 +94,19 @@ function onSelectionModeDoubleClick(e: any) {
 
 function onSelectionModeContextMenu(e: any) {
     e.preventDefault();
+    contextMenuTopLeftPosition = getTargetRowColSelectionMode(e);
 
     if (tableSelectionContextMenuElement) {
         tableSelectionContextMenuElement.style.display = 'block';
         tableSelectionContextMenuElement.style.left = e.clientX + 'px';
         tableSelectionContextMenuElement.style.top = e.clientY + 'px';
+    }
+
+    const queryResult = getQueryResult();
+    if (queryResult?.fields[contextMenuTopLeftPosition.col].flags.includes('notnull')) {
+        tableSelectionContextMenuSetNullElement.classList.add('disabled');
+    } else {
+        tableSelectionContextMenuSetNullElement.classList.remove('disabled');
     }
 }
 
@@ -111,6 +130,10 @@ function selectionModeClearSelectedCols() {
 }
 
 function onSelectionModeMouseUp(e: any) {
+    // right click for context menu
+    if (e.button === 2) {
+        return;
+    }
     isSelectionModeMouseDown = false;
 
     const { row, col } = getTargetRowColSelectionMode(e);
@@ -161,9 +184,16 @@ export function addSelectedRowsToDeletion() {
 }
 
 export function copySelectedColumns() {
+    // if there is something selected, prefer the selection
+    const selection = window.getSelection()?.toString();
+    if (selection) {
+        copyToClipboard(selection);
+        return;
+    }
+
     const tableElement = getTableElement();
     const selectedColumns = tableElement?.querySelectorAll('div.col--selected');
-    if (selectedColumns === undefined) {
+    if (selectedColumns === undefined || selectedColumns.length === 0) {
         return; // noting to copy
     }
 
@@ -172,7 +202,9 @@ export function copySelectedColumns() {
         const col = selectedColumns[i];
         if (col.classList.contains('row-number')) {
             // remove extra separator at end of last column and add newline
-            content = content.slice(0, -1) + '\n';
+            if (i !== 0) {
+                content = content.slice(0, -1) + '\n';
+            }
             continue;
         }
         let value = col.innerHTML;
@@ -204,6 +236,64 @@ export function renderSelectionModeContextMenu() {
         }
 
         setQueryResultDeletions(getQueryResultDeletions().filter(rowNumber => !rows.includes(rowNumber)));
+        updateOnPushChangesState();
+    });
+
+    tableSelectionContextMenuSelectRowElement.addEventListener('click', (e) => {
+        const row = tableElement?.children[contextMenuTopLeftPosition.row];
+        if (!row) {
+            logError('Error selecting row (row not found)', row);
+            return;
+        }
+        for (let i = 0; i < row?.children.length || 0; i++) {
+            row.children[i].classList.add('col--selected');
+        }
+    });
+
+    tableSelectionContextMenuSelectColumnElement.addEventListener('click', (e) => {
+        if (!tableElement) {
+            logError('Error selecting all (tableElement not found)', tableElement);
+            return;
+        }
+        for (let i = 0; i < tableElement.children.length || 0; i++) {
+            const col = tableElement.children[i].children[contextMenuTopLeftPosition.col + 1];
+            col.classList.add('col--selected');
+        }
+    });
+
+    tableSelectionContextMenuSelectAllElement.addEventListener('click', (e) => {
+        if (!tableElement) {
+            logError('Error selecting all (tableElement not found)', tableElement);
+            return;
+        }
+        for (let i = 0; i < tableElement.children.length || 0; i++) {
+            const row = tableElement.children[i];
+            for (let i = 0; i < row.children.length || 0; i++) {
+                row.children[i].classList.add('col--selected');
+            }
+        }
+
+    });
+
+    tableSelectionContextMenuDeleteRowElement.addEventListener('click', (e) => {
+        tableElement?.children[contextMenuTopLeftPosition.row].classList.add('row--delete');
+        setQueryResultDeletions([...getQueryResultDeletions(), contextMenuTopLeftPosition.row]);
+        updateOnPushChangesState();
+    });
+
+    tableSelectionContextMenuDeleteRowClearElement.addEventListener('click', (e) => {
+        tableElement?.children[contextMenuTopLeftPosition.row].classList.remove('row--delete');
+        setQueryResultDeletions(getQueryResultDeletions().filter(rowNumber => rowNumber !== contextMenuTopLeftPosition.row));
+        updateOnPushChangesState();
+    });
+
+    tableSelectionContextMenuSetNullElement.addEventListener('click', (e) => {
+        const target = tableElement?.children[contextMenuTopLeftPosition.row].children[contextMenuTopLeftPosition.col + 1];
+        if (target) {
+            target.setAttribute('data-placeholder', '<NULL>');
+            target.innerHTML = '';
+        }
+        updateDataChanges(contextMenuTopLeftPosition.row, contextMenuTopLeftPosition.col, null);
         updateOnPushChangesState();
     });